+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/installer/session/helpers/UriHelpers.kt
@@ -22,39 +22,7 @@ import android.content.res.AssetFileDescriptor
 import android.net.Uri
 import android.os.Build
 import android.os.CancellationSignal
-import android.os.Process
 import android.provider.OpenableColumns
-import java.io.File
-import java.io.FileNotFoundException
-
-@JvmSynthetic
-internal fun Uri.toFile(context: Context, signal: CancellationSignal): File {
-	if (scheme == ContentResolver.SCHEME_FILE) {
-		return File(requireNotNull(path) { "Uri path is null: $this" })
-	}
-	try {
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-			context.contentResolver.openFileDescriptor(this, "r", signal)
-		} else {
-			context.contentResolver.openFileDescriptor(this, "r")
-		}.use { fileDescriptor ->
-			if (fileDescriptor == null) {
-				throw NullPointerException("ParcelFileDescriptor was null: $this")
-			}
-			val path = "/proc/${Process.myPid()}/fd/${fileDescriptor.fd}"
-			val canonicalPath = File(path).canonicalPath.let { canonicalPath ->
-				if (canonicalPath.startsWith("/mnt/media_rw")) {
-					canonicalPath.replaceFirst("/mnt/media_rw", "/storage")
-				} else {
-					canonicalPath
-				}
-			}
-			return File(canonicalPath)
-		}
-	} catch (_: FileNotFoundException) {
-		return File("")
-	}
-}
 
 @JvmSynthetic
 internal fun Context.openAssetFileDescriptor(uri: Uri, signal: CancellationSignal): AssetFileDescriptor? {