+++ ackpine-runtime/src/main/kotlin/ru/solrudev/ackpine/AckpineFileProvider.kt
@@ -16,7 +16,22 @@
 
 package ru.solrudev.ackpine
 
+import android.content.Context
+import android.content.pm.ProviderInfo
+import androidx.annotation.RestrictTo
 import androidx.core.content.FileProvider
-import ru.solrudev.ackpine.core.R
+import ru.solrudev.ackpine.runtime.R
 
-internal class AckpineFileProvider : FileProvider(R.xml.file_provider_paths)
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public class AckpineFileProvider : FileProvider(R.xml.ackpine_file_provider_paths) {
+
+	override fun attachInfo(context: Context, info: ProviderInfo) {
+		super.attachInfo(context, info)
+		authority = info.authority
+	}
+
+	public companion object {
+		public lateinit var authority: String
+			private set
+	}
+}