+++ ackpine-runtime/src/main/kotlin/ru/solrudev/ackpine/helpers/UriHelpers.kt
@@ -19,19 +19,24 @@ package ru.solrudev.ackpine.helpers
 import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
+import android.os.Build
 import android.os.CancellationSignal
 import android.os.Process
-import android.provider.OpenableColumns
+import androidx.annotation.RestrictTo
 import java.io.File
 import java.io.FileNotFoundException
 
-@JvmSynthetic
-internal fun Uri.toFile(context: Context, signal: CancellationSignal? = null): File {
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public fun Uri.toFile(context: Context, signal: CancellationSignal? = null): File {
 	if (scheme == ContentResolver.SCHEME_FILE) {
 		return File(requireNotNull(path) { "Uri path is null: $this" })
 	}
 	try {
-		context.contentResolver.openFileDescriptor(this, "r", signal).use { fileDescriptor ->
+		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+			context.contentResolver.openFileDescriptor(this, "r", signal)
+		} else {
+			context.contentResolver.openFileDescriptor(this, "r")
+		}.use { fileDescriptor ->
 			if (fileDescriptor == null) {
 				throw NullPointerException("ParcelFileDescriptor was null: $this")
 			}
@@ -47,20 +52,5 @@ internal fun Uri.toFile(context: Context, signal: CancellationSignal? = null): F
 		}
 	} catch (_: FileNotFoundException) {
 		return File("")
-	}
-}
-
-@JvmSynthetic
-internal fun Uri.displayNameAndSize(context: Context): Pair<String, Long> {
-	context.contentResolver.query(
-		this,
-		arrayOf(OpenableColumns.DISPLAY_NAME, OpenableColumns.SIZE),
-		null, null, null
-	).use { cursor ->
-		cursor ?: return "" to -1L
-		if (!cursor.moveToFirst()) {
-			return "" to -1L
-		}
-		return cursor.getString(0).orEmpty() to cursor.getLong(1)
 	}
 }