+++ ackpine-splits/src/main/kotlin/ru/solrudev/ackpine/splits/helpers/IsApk.kt
@@ -14,14 +14,22 @@
  * limitations under the License.
  */
 
-package ru.solrudev.ackpine.helpers
+package ru.solrudev.ackpine.splits.helpers
 
-private const val CONFIG_PART = "config."
+import android.content.Context
+import android.net.Uri
+import java.io.File
+import java.util.zip.ZipEntry
+
+@get:JvmSynthetic
+internal val ZipEntry.isApk: Boolean
+	get() = name.endsWith(".apk", ignoreCase = true) && !isDirectory
+
+@get:JvmSynthetic
+internal val File.isApk: Boolean
+	get() = name.endsWith(".apk", ignoreCase = true) && !isDirectory
 
 @JvmSynthetic
-internal fun splitTypePart(name: String): String? {
-	if (!name.contains(CONFIG_PART, ignoreCase = true) && !name.contains(".$CONFIG_PART", ignoreCase = true)) {
-		return null
-	}
-	return name.substringAfter(CONFIG_PART).lowercase()
+internal fun Uri.isApk(context: Context): Boolean {
+	return context.contentResolver.getType(this) == "application/vnd.android.package-archive"
 }